reorganizing code into more folders
