+++ src/qp/core/pdf_gen.py
