+++ src/qp/parameterizations/sparse_rep.py
